// bios/src/client/interfaces/IRelatedDepartmentsWorkPlaces.ts
import { IRoot } from "./main";

namespace IRelatedDepartmentsWorkPlaces {
    export interface IBase extends IRoot {
        id: string;
        workplace_id: string;
        workplace_record_id: string;
        related_department_id: string;
        related_department_record_id: string;
        is_active: boolean;
        is_deleted: boolean;
    }
    export interface ICreate {
        id: string;
        workplace_id: string;
        workplace_record_id: string;
        related_department_id: string;
        related_department_record_id: string;
    }
}

export default IRelatedDepartmentsWorkPlaces;